refactor: move time-related system calls to the same file
